+++ protocol/client/frontend/virtual/interface.h
@@ -36,8 +36,8 @@ typedef struct VirtualInterface {
         void (*destroy_window)(int window_id);
     } lifecycle;
     struct {
-        void (*set_callback)(void (*cb)(unsigned int level, const char* line, int session_id));
-        void (*set_session_id)(int session_id);
+        void (*set_callback)(void (*cb)(unsigned int level, const char* line, long session_id));
+        void (*set_session_id)(long session_id);
         int (*get_session_id)(void);
     } logging;
     struct {
