+++ protocol/whist/logging/logging.h
@@ -301,10 +301,19 @@ void test_set_pause_state_on_logger_thread(bool pause);
  * @param cb  The callback, which is a void-returning function taking
  *            a log level and the logged line, as a string.
  */
-void whist_log_set_external_logger_callback(void (*cb)(unsigned int level, const char* line, int session_id));
+void whist_log_set_external_logger_callback(void (*cb)(unsigned int level, const char* line, long session_id));
 
-void whist_set_session_id(int session_id);
+/**
+ * Sets the session ID
+ *
+ * @param session_id  The session ID
+ */
+void whist_set_session_id(long session_id);
 
+/**
+ * Returns the session ID
+ *
+ */
 int whist_get_session_id(void);
 
 #endif  // LOGGING_H
